[bugfix] fix usage of atomic.AddUint32
Assigning the result of `atomic.AddUint32` back to the variable is not
atomic. The return value is to get the result of the operation after
executing it atomically. `req.Seq` should be read with `LoadUint32`
exclusively if we want the latest value. Otherwise (like here), we
should use the returned value (into another variable).

Go 1.18 now detect such an assignemt and turns it into an error.